+++ config/hypr/lua/user_overrides.lua
@@ -13,7 +13,6 @@ local files = {
   "system_env.lua",
   "system_startup.lua",
   "system_window_rules.lua",
-  "system_layer_rules.lua",
   "system_keybinds.lua",
   "system_settings.lua",
   "system_laptops.lua",